Ops_Tech_Builder_Test/requirements.txt
# re is from Python standard library, no need to install

flask
fastapi
pandas
openpyxl
requests
openai
uvicorn
google-generativeai
